backend/simpletesting/requirements.txt
asgiref==3.6.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.14.0
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environ==1.0
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
httpie==3.2.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
PySocks==1.7.1
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
